refactor: 컨벤션 통일성,(Request-요청, Response-반환)
